// src/models/repositorio.ts
export class Repositorio {

  id: number;
  aseguradora: number;
  canal: number;
  usuario: string;
  dominio: string;
  codigo: string;
  tipo_vehiculo: string;
  marca: string;
  modelo: string;
  anio: number;
  infoauto: number;
  color: string;
  km: number;
  combustible: string;
  interiores: string;
  uso: string;
  chasis: string;
  motor: string;
  nombre: string;
  apellido: string;
  dni: string;
  telefono: string;
  celular: string;
  email: string;
  genero: string;
  cp: number;
  provincia: number;
  localidad: number;
  calle: string;
  numero: string;
  depto: string;
  entre_calles: string;
  prod_nombre: string;
  prod_apellido: string;
  prod_email: string;
  prod_telefono: string;
  prod_cuit: string;
  estado: number;
  id_brb: number;
  enc_id_brb: string;
  gestionador_email: string;
  gestionador_cuit: string;
  organizador_cuit: string;
  localidad_text: string;
  provincia_text: string;
  key: string;
  documentacion: string;
  prestador_rastreo: string;
  fecha_envio_email: string;
  fecha_recepcion_email: string;
  cobertura: string;
  documentos_faltantes: string;
  tipo: number;
  franquicia: number;
  vencimiento_gnc: string;
  envia_ws: string;
  agenda: Date;
  observaciones: string;
  user_brb: number;
  anula_fecha: Date;
  anula_usuario: number;
  lo_audita: number;
  envio_link: string;
  fecha_envio_link: Date;
  call_back: string;
  smtp: string;
  organizador: string;
  reinspeccionada: boolean;
  precio_ia: number;
  fecha_envio_aviso: Date;
  delegacion: string;
  zona_riesgo: string;
  ejecutivo_cia: string;
  reinspeccion: number;
  cant_email: number;
  cant_sms: number;
  user_ws: number;
  auto_clasico: number;
  actualiza_usuario: number;
  opt: string;
  grupo: string;


  toJSON(extra?: {}) {
    return Object.assign({
      id: this.id,
      aseguradora: this.aseguradora,
      canal: this.canal,
      usuario: this.usuario,
      dominio: this.dominio,
      codigo: this.codigo,
      tipo_vehiculo: this.tipo_vehiculo,
      marca: this.marca,
      modelo: this.modelo,
      anio: this.anio,
      infoauto: this.infoauto,
      color: this.color,
      km: this.km,
      combustible: this.combustible,
      interiores: this.interiores,
      uso: this.uso,
      chasis: this.chasis,
      motor: this.motor,
      nombre: this.nombre,
      apellido: this.apellido,
      dni: this.dni,
      telefono: this.telefono,
      celular: this.celular,
      email: this.email,
      genero: this.genero,
      cp: this.cp,
      provincia: this.provincia,
      localidad: this.localidad,
      calle: this.calle,
      numero: this.numero,
      depto: this.depto,
      entre_calles: this.entre_calles,
      prod_nombre: this.prod_nombre,
      prod_apellido: this.prod_apellido,
      prod_email: this.prod_email,
      prod_telefono: this.prod_telefono,
      prod_cuit: this.prod_cuit,
      estado: this.estado,
      id_brb: this.id_brb,
      gestionador_email: this.gestionador_email,
      gestionador_cuit: this.gestionador_cuit,
      organizador_cuit: this.organizador_cuit,
      localidad_text: this.localidad_text,
      provincia_text: this.provincia_text,
      key: this.key,
      documentacion: this.documentacion,
      prestador_rastreo: this.prestador_rastreo,
      fecha_envio_email: this.fecha_envio_email,
      fecha_recepcion_email: this.fecha_recepcion_email,
      cobertura: this.cobertura,
      documentos_faltantes: this.documentos_faltantes,
      tipo: this.tipo,
      franquicia: this.franquicia,
      vencimiento_gnc: this.vencimiento_gnc,
      envia_ws: this.envia_ws,
      agenda: this.agenda,
      observaciones: this.observaciones,
      user_brb: this.user_brb,
      anula_fecha: this.anula_fecha,
      anula_usuario: this.anula_usuario,
      lo_audita: this.lo_audita,
      envio_link: this.envio_link,
      fecha_envio_link: this.fecha_envio_link,
      call_back: this.call_back,
      smtp: this.smtp,
      organizador: this.organizador,
      reinspeccionada: this.reinspeccionada,
      precio_ia: this.precio_ia,
      fecha_envio_aviso: this.fecha_envio_aviso,
      delegacion: this.delegacion,
      zona_riesgo: this.zona_riesgo,
      ejecutivo_cia: this.ejecutivo_cia,
      reinspeccion: this.reinspeccion,
      cant_email: this.cant_email,
      cant_sms: this.cant_sms,
      user_ws: this.user_ws,
      auto_clasico: this.auto_clasico,
      actualiza_usuario: this.actualiza_usuario,
      opt: this.opt,
      grupo: this.grupo,
      enc_id_brb: this.enc_id_brb
    }, extra)
  }

}